12/14 コンポーネントの分割
