+++ src/components/home_lists/Item.jsx
@@ -1,0 +1,73 @@
+import styled from "styled-components";
+
+import img from "../../images/1221552_00_1670457676.webp"
+import str from "../../images/pic_star45.png"
+
+
+const ItemBox = styled.div`
+  width: 144px;
+  height: 100%;
+  margin: 0 15px 0 0;
+  padding: 0;
+`;
+const ItemSet = styled.div`
+  width: 144px;
+  margin: 10px 0 0 0;
+  padding: 0;
+`;
+const SetName = styled.p`
+  width: auto;
+  margin: 0;
+  padding: 0;
+  font-size: 13px;
+  font-weight: 700;
+`;
+const SetCity = styled.p`
+  width: auto;
+  margin: 3px 0 0 0;
+  padding: 0;
+  font-size: 10px;
+`;
+const SetShow = styled.p`
+  width: auto;
+  margin: 0;
+  padding: 0;
+`;
+const SetPrice = styled.p`
+  width: auto;
+  margin: 3px 0 0 0;
+  padding: 0;
+  font-size: 13px;
+`;
+const SetText = styled.p`
+  width: auto;
+  margin: 3px 0 0 0;
+  padding: 0;
+  font-size: 10px;
+`;
+const SetReview = styled.div`
+  width: auto;
+  margin: 3px 0 0 0;
+  padding: 0;
+`;
+
+
+const Item = () => {
+  return (
+    <ItemBox>
+      <img src={img} alt='img' height='144px' width='144px' />
+      <ItemSet>
+        <SetName>【先行受付】山梨県産シャインマスカ...</SetName>
+        <SetCity>山梨県山梨市</SetCity>
+        <SetShow>
+          <SetPrice>寄付金額</SetPrice>
+          <SetText>山梨県の豊かな自然で育った、味わい深いシャインマスカットです。</SetText>
+          <SetReview><img src={str} alt='str' height='13px' width='82px' /></SetReview>
+        </SetShow>
+      </ItemSet>
+    </ItemBox>
+  );
+};
+
+
+export default Item;
